purchasedIdList 추가
